+++ src/main/resources/com/dataintuitive/viash/helpers/bashutils/ViashDockerFuns.sh
@@ -1,4 +1,4 @@
-#### Helper functions for building/pulling Docker environments for viash
+######## Helper functions for setting up Docker images for viash ########
 
 
 # ViashDockerRemoteTagCheck: check whether a Docker image is available 
@@ -39,7 +39,7 @@ function ViashDockerLocalTagCheck {
 #   ViashDockerPull sdaizudceahifu
 #   echo $?                                     # returns '1'
 function ViashDockerPull {
-  docker pull $1
+  docker pull $1 && return 0 || return 1
 }
 
 # ViashDockerPullElseBuild: pull a Docker image, else build it
@@ -48,43 +48,57 @@ function ViashDockerPull {
 # examples:
 #   ViashDockerPullElseBuild mynewcomponent
 function ViashDockerPullElseBuild {
+  set +e
   ViashDockerPull $1
-  if [ $? -ne 0 ]; then
-    ViashDockerBuild $1
+  out=$?
+  set -e
+  if [ $out -ne 0 ]; then
+    ViashDockerBuild $@
   fi
 }
 
-# ViashDockerPullElseBuild: pull a Docker image, else build it
+# ViashDockerSetup: create a Docker image, according to specified docker setup strategy
 #
 # $1                  : image identifier with format `[registry/]image[:tag]`
+# $2                  : docker setup strategy, see DockerSetupStrategy.scala
 # ViashDockerBuild    : a Bash function which builds a docker image, takes image identifier as argument.
 # examples:
 #   ViashDockerPullElseBuild mynewcomponent alwaysbuild
 function ViashDockerSetup {
-  VSH_DOCKER_ID="$1"
-  VSH_DOCKER_STRATEGY="$2"
-  if [ "$VSH_DOCKER_STRATEGY" == "alwaysbuild" ]; then
-    ViashDockerBuild $VSH_DOCKER_ID
-  elif [ "$VSH_DOCKER_STRATEGY" == "alwayspull" ]; then
-    ViashDockerPull $VSH_DOCKER_ID
-  elif [ "$VSH_DOCKER_STRATEGY" == "alwayspullelsebuild" ]; then
-    ViashDockerPullElseBuild $VSH_DOCKER_ID
-  elif [ "$VSH_DOCKER_STRATEGY" == "dontbother" ]; then
-    :
-  elif [[ "$VSH_DOCKER_STRATEGY" =~ ^ifneedbe ]]; then
-    ViashDockerLocalTagCheck $VSH_DOCKER_ID
+  VSHD_ID="$1"
+  VSHD_STRAT="$2"
+  if [ "$VSHD_STRAT" == "alwaysbuild" -o "$VSHD_STRAT" == "build" ]; then
+    ViashDockerBuild $VSHD_ID --no-cache
+  elif [ "$VSHD_STRAT" == "alwayspull" -o "$VSHD_STRAT" == "pull" ]; then
+    ViashDockerPull $VSHD_ID
+  elif [ "$VSHD_STRAT" == "alwayspullelsebuild" -o "$VSHD_STRAT" == "pullelsebuild" ]; then
+    ViashDockerPullElseBuild $VSHD_ID --no-cache
+  elif [ "$VSHD_STRAT" == "alwayspullelsecachedbuild" -o "$VSHD_STRAT" == "pullelsecachedbuild" ]; then
+    ViashDockerPullElseBuild $VSHD_ID
+  elif [ "$VSHD_STRAT" == "alwayscachedbuild" -o "$VSHD_STRAT" == "cachedbuild" ]; then
+    ViashDockerBuild $VSHD_ID
+  elif [ "$VSHD_STRAT" == "donothing" -o "$VSHD_STRAT" == "meh" ]; then
+    echo "Skipping setup."
+  elif [[ "$VSHD_STRAT" =~ ^ifneedbe ]]; then
+    ViashDockerLocalTagCheck $VSHD_ID
     if [ $? -eq 0 ]; then
-      echo "Image $VSH_DOCKER_ID already exists"
-    elif [ "$VSH_DOCKER_STRATEGY" == "ifneedbebuild" ]; then
-      ViashDockerBuild $VSH_DOCKER_ID
-    elif [ "$VSH_DOCKER_STRATEGY" == "ifneedbepull" ]; then
-      ViashDockerPull $VSH_DOCKER_ID
-    elif [ "$VSH_DOCKER_STRATEGY" == "ifneedbepullelsebuild" ]; then
-      ViashDockerPullElseBuild $VSH_DOCKER_ID
+      echo "Image $VSHD_ID already exists"
+    elif [ "$VSHD_STRAT" == "ifneedbebuild" ]; then
+      ViashDockerBuild $VSHD_ID --no-cache
+    elif [ "$VSHD_STRAT" == "ifneedbecachedbuild" ]; then
+      ViashDockerBuild $VSHD_ID
+    elif [ "$VSHD_STRAT" == "ifneedbepull" ]; then
+      ViashDockerPull $VSHD_ID
+    elif [ "$VSHD_STRAT" == "ifneedbepullelsebuild" ]; then
+      ViashDockerPullElseBuild $VSHD_ID --no-cache
+    elif [ "$VSHD_STRAT" == "ifneedbepullelsecachedbuild" ]; then
+      ViashDockerPullElseBuild $VSHD_ID
     else
-      echo "Unrecognised Docker strategy: $VSH_DOCKER_STRATEGY"
+      echo "Unrecognised Docker strategy: $VSHD_STRAT"
     fi
   else
-    echo "Unrecognised Docker strategy: $VSH_DOCKER_STRATEGY"
+    echo "Unrecognised Docker strategy: $VSHD_STRAT"
   fi
 }
+
+######## End of helper functions for setting up Docker images for viash ########